Fixed queuing problems with ProxyPeer threads
The first problem was that nextChunk was being read in a different synchronization block than the queue, so they could be out of sync.

The second problem was that the 'chunk' command wasn't removed from the queue even if the pendingChunk list became empty (because another ProxyPeer downloaded the chunk). This caused a NoSuchElementException from the pendingChunk list.